+++ tests/testthat/test_twa.R
@@ -22,15 +22,9 @@ context("Calculation of maximum time weighted average concentrations (TWAs)")
 test_that("Time weighted average concentrations are correct", {
   skip_on_cran()
   twa_models <- c("SFO", "FOMC", "DFOP", "HS")
-  travis_env <- Sys.getenv("TRAVIS")
-  if (travis_env == "") {
-    travis <- FALSE
-  } else {
-    travis <- TRUE
-  }
   fits <- mmkin(twa_models,
                 list(FOCUS_C = FOCUS_2006_C, FOCUS_D = FOCUS_2006_D),
-                quiet = TRUE, cores = if (travis) 1 else 8)
+                quiet = TRUE, cores = if (Sys.getenv("TRAVIS") == "") 15 else 1)
 
   outtimes_10 <- seq(0, 10, length.out = 10000)